requirements.txt
boto3
requests>=2.20
Pillow==9.2.0
mypy-boto3-ssm
mypy-boto3-sns
mypy-boto3-s3
mypy-boto3-lambda
pytest
requests
pytest-bdd
pytest-splinter
awscli
awscli-local
pylint